+++ src/test/java/com/harry/market/MarketApplicationTests.java
@@ -49,14 +49,14 @@ class MarketApplicationTests {
             String price = el.getElementsByClass("p-price").get(0).getElementsByTag("i").get(0).text();
 
             Item item = new Item();
-            BigInteger nid = new BigInteger(id);
+            Long nid = new Long(id);
             item.setId(nid);
             item.setName(pname);
             item.setPhoto(img);
             BigDecimal nprice = new BigDecimal(price);
             item.setPrice(nprice);
             Random seed = new Random();
-            BigInteger sid = BigInteger.valueOf(seed.nextInt(10));
+            Long sid = Long.valueOf(seed.nextInt(10));
             item.setSeller_id(sid);
             item.setKind("日用家居");
             System.out.println(item);
